% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configGDAL.R
\name{install.gdal}
\alias{install.gdal}
\title{Download and configure GDAL binary utilities under windows.
This function help to download and install GDAL binary utilities.}
\usage{
install.gdal(install_path = "c:/gdal_map_win", proxy_url = NULL)
}
\arguments{
\item{install_path}{character. path for unzip of GDAL binary utilities}

\item{proxy_url}{character indicating proxy url. The format is 'http://server_host:port'}
}
\value{
returns an logical value which represents whether GDAL binary installing is success
}
\description{
Download and configure GDAL binary utilities under windows.
This function help to download and install GDAL binary utilities.
}
\details{
This function will download gdal binary(with proxy), and set the search path for gdal_setInstallation automatically.
}
\examples{
install.gdal(install_path='/c/gdalwin')
}

