% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGRID.R
\name{loadGRID}
\alias{loadGRID}
\title{Generate a grid polygon of all daily rozone dataset in a single year.}
\usage{
loadGRID(datapath = "data-raw/2010")
}
\arguments{
\item{datapath}{the path of year folder}
}
\value{
a spatial polygon data frame which contains daily datasets in this year.
}
\description{
This function generatea a grid-shape polygons of  all daily rozone dataset in one year.
}
\examples{
grid.poly<-loadGRID()
plot(grid.poly)
grid.poly<-loadGRID(datapath='data-raw/2010')
}

