% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize.R
\name{rasterize.gdal}
\alias{rasterize.gdal}
\title{High performance rasterize function using gdalUtils.}
\usage{
rasterize.gdal(x, field, res = 10000, extent = NULL)
}
\arguments{
\item{x}{a Spatial*DataFrame objects.}

\item{field}{character.The column name of the variable to be transferred}

\item{res}{numeric.Resolution of the output raster}

\item{extent}{extent object.Spatial extent of output raster.If missing, the extent of x will be used.}
}
\value{
RasterLayer
}
\description{
This function rasterize a vector data by gdal binary utils.
}
\examples{
###### the default path is 'data-raw/2010'. You should change it to your data folder. ###
grid.poly<-loadGRID()
plot(grid.poly)
###### set the path of data #######
grid.poly<-loadGRID(datapath='data-raw/2010')
}

