% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configGDAL.R
\name{install.gdal}
\alias{install.gdal}
\title{Configure GDAL binary utilities under windows.}
\usage{
install.gdal(install_path = NULL, proxy_url = NULL)
}
\arguments{
\item{install_path}{character. path for unzip of GDAL binary utilities}

\item{proxy_url}{character indicating proxy url. The format is 'http://server_host:port'}

\item{extent}{extent object.Spatial extent of output raster.If missing, the extent of x will be used.}
}
\description{
This function help to find without proxy or install GDAL binary utilities with proxy.
}
\details{
If proxy_url is avaliable, this function will download gdal binary with proxy, and set the search path for gdal_setInstallation automatically.

If proxy_url is missing, this function will open the download page with browser because author can't find a good netdisk to host these files.The only way to download them quickly is to download files by yourself.
}
\examples{
install.gdal(install_path='/tmp/gisinter',proxy_url="http://127.0.0.1:8087/")
}

