% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalstats.R
\name{zonalStat}
\alias{zonalStat}
\title{High performance zonal statistics function.}
\usage{
zonalStat(zonal.ras, val.ras, stat = "mean")
}
\arguments{
\item{zonal.ras}{a raster objects indicating each zone}

\item{val.ras}{the raster dataset to be calculated}

\item{stat}{character.statistics method to be performed}

\item{extent}{extent object.Spatial extent of output raster.If missing, the extent of x will be used.}
}
\value{
RasterLayer
}
\description{
This function calculate statistics for each zone defined by a zone dataset.
}
\examples{
###############################
# rasterize points
###############################
r <- raster(ncols=36, nrows=18)
n <- 1000
x <- runif(n) * 360 - 180
y <- runif(n) * 180 - 90
xy <- cbind(x, y)
# with a SpatialPointsDataFrame
vals <- 1:n
p <- data.frame(xy, name=vals)
coordinates(p) <- ~x+y
## If you are using windows,you must set local GDAL installation options before running function###
#Not run:
#gdal_setInstallation(search_path = 'the path of OSGeo4W64/bin',rescan = T)
r2 <- rasterize.gdal(p, field = 'name', res =res(r)[1],extent=extent(r))
}

